Changes to constants.py
