+++ constants.py
@@ -1,3 +1,5 @@
+from trip.helpers import concat_sort_city_tuples
+
 DRINK_PREF_CHOICES = [
     ("Frequently", "Frequently"),
     ("Socially", "Socially"),
@@ -139,14 +141,6 @@ ITALY_CITIES = [
     ("Rome", "Rome"),
     ("Venice", "Venice"),
 ]
-
-
-def concat_sort_city_tuples(*args):
-    return_list = []
-    for arg in args:
-        return_list += arg
-    return sorted(return_list)
-
 
 CITY_CHOICES = concat_sort_city_tuples(
     INDIAN_CITIES,
